backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Werkzeug==3.0.0
SQLAlchemy==2.0.16
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
bcrypt==4.1.2
cloudinary==1.33.0
Pillow==10.2.0
flask-cors==4.0.0 